Update and refactor frontend for SDK 0.8.8
