+++ frontend/src/components/AnalysisReport/index.tsx
@@ -1,3 +1,5 @@
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-nocheck
 import React, { useState } from "react";
 import {
   ResultFineGrainedParsed,
