+++ frontend/src/components/SystemsTable/index.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import "./index.css";
 import { Space } from "antd";
-import { SystemSubmitDrawer } from "../../components";
+import { SystemSubmitDrawer, AnalysisDrawer } from "../../components";
 import { backendClient } from "../../clients";
 import { SystemModel, newSystemModel } from "../../models/SystemModel";
 import { Filter, SystemTableTools } from "./SystemTableTools";
@@ -38,6 +38,7 @@ export function SystemsTable({ initialTaskFilter }: Props) {
 
   // systems selected in the table
   const [selectedSystemIDs, setSelectedSystemIDs] = useState<string[]>([]);
+
   // systems to be analyzed
   const [activeSystemIDs, setActiveSystemIDs] = useState<string[]>([]);
 
@@ -45,7 +46,9 @@ export function SystemsTable({ initialTaskFilter }: Props) {
   function getMetricsNames() {
     const metricNames = new Set<string>();
     for (const sys of systems) {
-      sys.analysis.getMetricNames().forEach((name) => metricNames.add(name));
+      Object.keys(sys.system_info.results.overall).forEach((name) =>
+        metricNames.add(name)
+      );
     }
     // if a task is selected, add all supported metrics to the options list
     if (taskFilter) {
@@ -142,6 +145,11 @@ export function SystemsTable({ initialTaskFilter }: Props) {
         metricNames={metricNames}
         selectedSystemIDs={selectedSystemIDs}
         setSelectedSystemIDs={setSelectedSystemIDs}
+        setActiveSystemIDs={setActiveSystemIDs}
+      />
+      <AnalysisDrawer
+        visible={activeSystemIDs.length !== 0}
+        systems={systems}
         activeSystemIDs={activeSystemIDs}
         setActiveSystemIDs={setActiveSystemIDs}
       />
