+++ frontend/src/components/index.ts
@@ -1,7 +1,7 @@
 export * from "./Layout";
 export * from "./SystemSubmitDrawer";
 export * from "./SystemsTable";
-export * from "./AnalysisReport";
+export * from "./AnalysisDrawer";
 export * from "./BarChart";
 export * from "./AnalysisTable";
 export * from "./TaskSelect";
